improve performance when exporting CSV tiles
The CSV export logic in 1.5.1+ appends to the output incrementally, but that's inefficient for large strings since they'll be duplicated in memory for each append (e.g. a map exported almost instantly in 1.4 can several seconds in 1.5.1). This commit adds a StringBuilder, so the full string is only generated when it's done appending.